+++ src/components/CreateList/CreateList.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Modal from 'react-modal';
 import PropTypes from 'prop-types';
-import '../styles/components/CreateList.scss';
+import './CreateList.scss';
 
 const CreateList = ({ createListIsOpen, handleCreateListClose, handleSubmitNewList }) => (
   <Modal
