+++ src/components/CreateList/index.js
@@ -1,0 +1,3 @@
+import CreateList from './CreateList';
+
+export default CreateList;
