+++ src/components/Header/Header.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../styles/components/Header.scss';
+import './Header.scss';
 import { Link } from 'react-router-dom';
 
 const Header = () => (
