+++ src/components/InsertItem/InsertItem.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Modal from 'react-modal';
 import PropTypes from 'prop-types';
-import '../styles/components/InsertItem.scss';
+import './InsertItem.scss';
 
 const InsertItem = ({
   isOpened,
