+++ src/components/InsertItem/index.js
@@ -1,0 +1,3 @@
+import InsertItem from './InsertItem';
+
+export default InsertItem;
