+++ src/components/Item/Item.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import '../styles/components/Item.scss';
+import './Item.scss';
 
 const Item = ({ item, addItemWishlist }) => {
   const { id, volumeInfo } = item;
