+++ src/components/Item/index.js
@@ -1,0 +1,3 @@
+import Item from './Item';
+
+export default Item;
