+++ src/components/Items/index.js
@@ -1,0 +1,3 @@
+import Items from './Items';
+
+export default Items;
