+++ src/components/Search/index.js
@@ -1,0 +1,3 @@
+import Search from './Search';
+
+export default Search;
