+++ src/components/SelectList/index.js
@@ -1,0 +1,3 @@
+import SelectList from './SelectList';
+
+export default SelectList;
