+++ src/components/WishList/WishList.jsx
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import '../styles/components/WishList.scss';
-import Item from './Item';
+import './WishList.scss';
+import Item from '../Item';
 
 class WishList extends Component {
   render() {
