+++ src/components/WishList/WishList.scss
