+++ src/components/WishList/index.js
@@ -1,0 +1,3 @@
+import WishList from './WishList';
+
+export default WishList;
