+++ src/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './styles/index.scss';
+import './index.scss';
 import App from './components/App';
 
 ReactDOM.render(<App />, document.getElementById('root'));
